backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.1
python-decouple>=3.8
PyJWT>=2.8.0
django-cors-headers>=4.3.1
drf-yasg>=1.21.7
